Aggiunti test sulla repo Subscription e finiti test sul layer repository

+++ user-subscription/src/main/java/com/example/usersubscription/repositories/SubscriptionRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SubscriptionRepo extends JpaRepository <Subscription,Integer> {
+public interface SubscriptionRepo extends JpaRepository <Subscription,Long> {
 
     List<Subscription> findByUser(User user);
